Fix TrailingSemicolon linter for deeply nested `@include`s
A previous commit, 54c5b7fb212, fixed a problem with the
TrailingSemicolon linter reporting lints on mixins that took a block.
While veryfing the behavior of that commit, @Steffen185 found another
issue with mixins with a deeper structure. He gave this repro case:

  $ cat test.scss
  @include foo {
    body {
      background: #f00;
    }
  }
  $ scss-lint test.scss
  test.scss:1 [W] TrailingSemicolon: Declaration should not have a space
  before the terminating semicolon

The fix for this issue involved including more node types in the
`has_nested_properties?`. Through some trial-and-error, I found that
adding the `Sass::Tree::RuleNode` type will do the right thing.

This commit closes #198.

Change-Id: I8f642f915a2b62c13e886425b10472bb1065de6c
Reviewed-on: http://gerrit.causes.com/40990
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Henric Trotzig <henric.trotzig@brigade.com>