Don't lint interpolated strings with StringQuotes
The `StringQuotes` linter would incorrectly report lints for strings
with interpolation, e.g.

    "<a href='#{$variable}'></a>"

In this case, the linter can't statically determine what kind of quote
character is the appropriate one, since the interpolation could change
the necessity for one type or the other.

Thus we disable linting for interpolated strings (but still lint the
expressions inside the interpolation).

Change-Id: I8db1108e927ee304481e2bbcaed65dd7029f26ce
Reviewed-on: http://gerrit.causes.com/42245
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>