Prefix all syntax error messages
Issue #232 requested that `scss-lint` give some indication of whether an
error was syntax- or style-related so that it would be displayed
appropriately by the Syntastic scss-lint plugin.

Since the only syntax error will ever be one reported by the `Runner`
class when parsing the document fails, standardize the error message to
being with "Syntax Error", so the scss-lint plugin can simply check the
error message to determine whether it is style or syntax.

Change-Id: I90183af30adba900ebcf94f688b8991ddcc1b663
Reviewed-on: http://gerrit.causes.com/44340
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>