Move RuboCop version pinning into Gemfile
We were declaring the pinned version of RuboCop as a
development dependency in the gemspec. This turns out to not be
technically correct, as it's not really a development dependency, but a
testing one (i.e. we don't want to enforce this requirement for
developers using this gem when developing other gems that require this
gem).

It turns out a way get around this is to declare the pinned version in
the `Gemfile`.

Change-Id: Ifa27a9c2cec89297300fc99f0aae483839eb636c
Reviewed-on: http://gerrit.causes.com/44724
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>