Don't call visit_selector on newline strings
When visiting a multi-line comma sequence, we would call
`visit_selector` on `String` objects since `Sequence`s intersperse
strings containing newlines in between each of the `SimpleSequence`s.

This didn't have any effect because the `String` class doesn't define
any `visit_*` methods, but it's a good idea to catch it earlier in the
stack.

Change-Id: Id09e37f66ed9224107bdac15131d2008ed03464f
Reviewed-on: http://gerrit.causes.com/45816
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>