Fix control comments with @include blocks
The `ControlCommentProcessor` would not mark lines in `@include` blocks
due to their children not being sorted by line number, so the last child
would not be the child on the lowest line in the file. This is due to
how we monkey patch the parse tree node's `children` method to return
additional `Sass::Script::Tree::Node`s when it makes sense to.

Fix this by finding the child with the lowest line by iterating over all
children. This should also handle cases we haven't discovered yet.

Change-Id: I4cb5007cb2096d919b72a58f3da0a1181524ef35
Reviewed-on: http://gerrit.causes.com/46076
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>