Add `break-{after,before,inside}` to list of properties
When you add `break-inside` property to your SCSS files it returns:

> PropertySpelling: Unknown property break-inside

While it's totally valid according to MDN:
https://developer.mozilla.org/en-US/docs/Web/CSS/break-inside

This commit adds the missing `break-after`, `break-before`, and
`break-inside` properties to the list of valid properties. These were
missing because we had been using the list of properties at
https://www.chromestatus.com/metrics/css/popularity but it did not
include this property because it is not supported by Chrome. There may
be other properties that we are similarly missing because of this data
source.

Fixes #356.

Change-Id: Ic0dc8cc7d98e95e709bc8a0bd2c12edc2acd845a
Reviewed-on: http://gerrit.causes.com/46079
Reviewed-by: Joe Lencioni <joe.lencioni@brigade.com>
Tested-by: jenkins <jenkins@brigade.com>