Fix recognition of control comments in selectors
When writing a control comment in a selector, e.g.

    .buttonText, // scss-lint:disable SelectorFormat
    .a-class {
      ...
    }

...that control comment would not get processed.

This is due to the Sass parser not parsing comments within these
selectors as `CommentNodes`, but simply includes the string itself in
the selector.

We work around this by explicitly searching the selector for comments
via regex, and if so apply the comment to the entire selector.

This isn't ideal since a comment on a different line could still disable
a lint on another line. This is because Sass also reports the line of
the entire selector as the beginning of the rule set, even if it spans
multiple lines.

To get around this will require some work, so we implement a solution
that gets us some of the way there and will probably work fine for most
cases.

Change-Id: I6c6499dca3f8693a5c3dc0bf007ae882826ba067
Reviewed-on: http://gerrit.causes.com/46089
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>