Remove == method from Config class
The presence of this method was breaking the build with the release of
RSpec 3.2.

Remove it and change the relevant specs to compare the `options` hash of
the config directly.

This is obviously not ideal, but I lack the time to figure out why the
original error was occurring, and this isn't used by any external
consumers so it's a minor offense.

Change-Id: If72d4b67df60875d95687778dcb63dcd36c0b677
Reviewed-on: http://gerrit.causes.com/46287
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>