Remove inherit_from and per-directory configuration support
These two features were created very early in `scss-lint`'s history, and
were largely a result of trying to emulating the `rubocop` project's
configuration system.

Over time, it's become clear that the configuration system is the most
complex part of `scss-lint`, both in implementation as well as tests. At
the core of this complexity is the ability to "inherit" configuration
files via `inherit_from`, and to load directory-specific configurations
based on the location of an SCSS file.

I think supporting these features encourages messier and harder to
understand configurations. After seeing `scss-lint` used in the wild,
it's clear that a majority of users only have an `.scss-lint.yml` file
in the root-level directory of their repositories, and furthermore those
who actually have custom configurations in subdirectories could easily
merge those configurations into the root config.

By allowing configurations to be located anywhere in the repository, the
developer now has to search multiple files in order to reconcile their
configuration, rather than consult a single file. While this may mean
the single file is larger, we're not talking thousands of lines of
configuration.

Removing these two features free us to simplify the configuration system
significantly, and opens up some other refactoring opportunities.

Change-Id: I22cd17629420ab9bf2d471830336fb2cdd66ad98
Reviewed-on: http://gerrit.causes.com/46307
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>