Fix handling of list literals with empty lists
The `BangFormat`, `ImportantRule`, and `VendorPrefixes` linters would
fail on SCSS like the following:

    p {
      content: 1 ();
    }

This is due to a bug in `Sass::Tree::Node` parser's `to_sass` method,
which each of these linters used. Fix the problem by switching to
inspecting the original source code via the `source_from_range` helper
so we don't rely on `to_sass`.

Change-Id: I836389461f45dfcfc6f9e86abf0843db8b2f628f
Reviewed-on: http://gerrit.causes.com/46920
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>