Actually fix allow_non_nested_indentation option
The previous commit solved some problems, but did not handle the case
where an arbitrarily nested ruleset contained incorrectly nested
non-rulesets, e.g.

    .one {
      color: #000;
    }
      .two { // Arbitrarily nested
          margin: 0;  // This is indented too much
      }

The solution is when in this special mode to compare the node's
indentation relative to its parent's indentation, rather than the
absolute expected indentation width.

The resulting code here is pretty disgusting, but that's a function of
the rather arbitrary set of rules needed to describe this indentation
behavior.

Change-Id: I390903604024de66732934973a8203054e88721c
Reviewed-on: http://gerrit.causes.com/48041
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>