Improve lint message of DeclarationOrder
Depending on the situation, the line number returned by this linter
would be confusing and not necessarily correspond with the actual node
that was out of place.

Improve the message returned to actually display the line numbers of the
nodes that need swapping in order to satisfy the expected sort order.