Display helpeful error when all matched files are excluded
Previously, when running `scss-lint` and all files specified via the
command-line or configuration were excluded by `--exclude` flags or
`exclude` files in configuration, it would return successfully saying no
lints were found.

This is a bit misleading, as it is likely that the user did not intend
to filter all files.

Change the behavior to report a helpful error message and custom exit
code so it can be detected in shell scripts and other integrations.