Add YARD to Shorthand linter
The types of arguments and return types of functions are helpful for
understanding a bit of code. Since I was in here, I decided to flesh
out the YARD for this file. I'm not sure that I got it all right, but it
seems at least mostly accurate to me.