Remove special case logic to handle all files filtered
This was added in 2d62be12 with the theory that users would appreciate
being warned when all files specified were excluded.

In practice, this has confused people and resulted in this special case
needing to be handled not just here, but in other tools like Overcommit
and the Vim syntastic checker. Remove it.