Fix TrailingSemicolon for variables with !default and !global
The TrailingSemicolon rule was not checking guarded or global variable
declarations properly. I determined that this is because we are using
`node.expr` for the source range, which in cases like:

  $foo: bar !default;

gives us a range that looks like:

  bar

which causes the linter to fail. Instead, if we just use the `node`
source range instead, we will get

  foo: bar !default;

so our checks work as expected. I considered always using `node` source
range, but I figured that if it is working for other cases that it would
be better to leave it as is.

Addresses #602