Remove Travis pre-commit hook
Travis now depends on net-http-persistent ~> 3 which requires Ruby >=
2.1.

While Ruby 2.0 was EOLed in February 2016, it's still the default Ruby
version to ship with Macs, so let's continue supporting it unless it's
too much of a hassle. Travis lint isn't a terribly important tool, so
remove it to remove the dependency.